Implemented biomass parabola into equilibrium tests.
Immediate todos:
- fix the stopping criterion-- right now I have an infinite loop
because it is looking for 0, which now is approached slowly and so
not achieved
- implement the parabola for the scenario tests. Note-- the
parabola does not yet have the warming sensitivity. It could be
added in case I would like to test warming sensitivity in the future.
But I likely won't, not for this context, this model, so better to
keep out unecessary features-- (which I believe is feature creap?)