Slight optimization to shooting loop: change inital SLR
This is the shooting loop, increasing SLR until it drowns the marsh, repeating
for increasing sediment concentrations.
I have made a small adjustment so that the initial SLR tried is the highest
in the previous sediement concentrations, assuming sediment concentrations are ordered
least to greatest.

This commit also has a high res run-- where the run is 400 years long and
we're testing .5 mm increments.